refactor(bundle-size): extract Mover/Groupper/Modalizer dummy factories
Each part used to inline its own per-instance create*DummyManager that
runs only when controlTab=false. With controlTab=true (the default), the
factory was dead code that still shipped — Terser couldn't prove it
unreachable since the gate is a runtime flag.

Move the three factories out to MoverDummyManager.ts /
GroupperDummyManager.ts / ModalizerDummyManager.ts. Register them on
tabsterCore.{mover,groupper,modalizer}DummyManagerFactory inside
getRootDummyInputs alongside the root factory; each part's constructor
calls the registered factory instead of an inline import. Without
getRootDummyInputs, the dummy code is absent from the bundle.

Also restore the phantom-dummy fallback for Root.moveOutWithDefaultAction
behind the same opt-in. With dummy inputs disabled the call no-ops
(consistent with the rest of the opt-in surface); when getRootDummyInputs
is called the registered moveOutOfRoot routes either through the root's
dummy manager or through DummyInputManager.moveWithPhantomDummy. This
restores the master behaviour for `controlTab: false` /
`rootDummyInputs: false` consumers without putting moveWithPhantomDummy
back into the always-on Root path.

Test setup now calls getRootDummyInputs unconditionally so
controlTab=false runs get the per-part dummy factories.

createTabster:                28.45 kB → 28.48 kB (+30 B for routing)
getMover:                     45.05 kB → 41.66 kB (-3.39 kB / -7.5%)
getGroupper:                  38.05 kB → 34.42 kB (-3.63 kB / -9.5%)
getModalizer:                 38.80 kB → 35.80 kB (-3.00 kB / -7.7%)
getModalizerWithDummyInputs:           40.19 kB (the controlTab=false
                                                  cost: +4.39 kB over
                                                  plain getModalizer)
